refactor: 수익에 대한 테스트 명확하게 리팩토링
